Revert "added animation to laser pointer"
This reverts commit fcbcd117e6ad15102774c71337e3800e43d10ef3.
