+++ Assets/Animations/MathewSantos.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 29fa5af4b4211da419836ba038442f36
-NativeFormatImporter:
+guid: 66f443c328ea34d46a2d8db608a8005b
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 7400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
