+++ Assets/Animations/MathewSantos/Laser.anim.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 53efe28cda6d73743a65aeed4c821d2c
+guid: ef071ca51f38ae84b87f9a2fac755709
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 9100000
+  mainObjectFileID: 7400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
