--- Assets/Animations/MattSantos.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e20f666f69623ad40928d12ee5dd3ba5
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
