+++ Assets/Scripts/Items/LaserPointer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f8ddab8829916524b8eb282fb73f5a9d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
